Adding equals and hashcode methods to PanacheEntity
